+++ registry/api/v2/descriptors.go
@@ -5,6 +5,7 @@ import (
 	"regexp"
 
 	"github.com/docker/distribution/digest"
+	"github.com/docker/distribution/registry/api/errcode"
 )
 
 var (
@@ -98,7 +99,7 @@ var (
 				Format:      "<length>",
 			},
 		},
-		ErrorCodes: []ErrorCode{
+		ErrorCodes: []errcode.ErrorCode{
 			ErrorCodeUnauthorized,
 		},
 		Body: BodyDescriptor{
@@ -119,7 +120,7 @@ var (
 				Format:      "<length>",
 			},
 		},
-		ErrorCodes: []ErrorCode{
+		ErrorCodes: []errcode.ErrorCode{
 			ErrorCodeUnauthorized,
 		},
 		Body: BodyDescriptor{
@@ -174,7 +175,7 @@ var APIDescriptor = struct {
 
 	// ErrorDescriptors provides a list of the error codes and their
 	// associated documentation and metadata.
-	ErrorDescriptors []ErrorDescriptor
+	ErrorDescriptors []errcode.ErrorDescriptor
 }{
 	RouteDescriptors: routeDescriptors,
 	ErrorDescriptors: errorDescriptors,
@@ -275,7 +276,7 @@ type ResponseDescriptor struct {
 
 	// ErrorCodes enumerates the error codes that may be returned along with
 	// the response.
-	ErrorCodes []ErrorCode
+	ErrorCodes []errcode.ErrorCode
 
 	// Body describes the body of the response, if any.
 	Body BodyDescriptor
@@ -317,30 +318,6 @@ type ParameterDescriptor struct {
 	Examples []string
 }
 
-// ErrorDescriptor provides relevant information about a given error code.
-type ErrorDescriptor struct {
-	// Code is the error code that this descriptor describes.
-	Code ErrorCode
-
-	// Value provides a unique, string key, often captilized with
-	// underscores, to identify the error code. This value is used as the
-	// keyed value when serializing api errors.
-	Value string
-
-	// Message is a short, human readable decription of the error condition
-	// included in API responses.
-	Message string
-
-	// Description provides a complete account of the errors purpose, suitable
-	// for use in documentation.
-	Description string
-
-	// HTTPStatusCodes provides a list of status under which this error
-	// condition may arise. If it is empty, the error condition may be seen
-	// for any status code.
-	HTTPStatusCodes []int
-}
-
 var routeDescriptors = []RouteDescriptor{
 	{
 		Name:        RouteNameBase,
@@ -374,7 +351,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 							},
@@ -438,7 +415,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameUnknown,
 								},
 							},
@@ -449,7 +426,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 							},
@@ -495,7 +472,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The name or reference was invalid.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeTagInvalid,
 								},
@@ -511,14 +488,14 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 							},
 							{
 								Description: "The named manifest is not known to the registry.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameUnknown,
 									ErrorCodeManifestUnknown,
 								},
@@ -573,7 +550,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeTagInvalid,
 									ErrorCodeManifestInvalid,
@@ -588,7 +565,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 							},
@@ -596,7 +573,7 @@ var routeDescriptors = []RouteDescriptor{
 								Name:        "Missing Layer(s)",
 								Description: "One or more layers may be missing during a manifest upload. If so, the missing layers will be enumerated in the error response.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUnknown,
 								},
 								Body: BodyDescriptor{
@@ -625,7 +602,7 @@ var routeDescriptors = []RouteDescriptor{
 										Format:      "<length>",
 									},
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 								Body: BodyDescriptor{
@@ -660,7 +637,7 @@ var routeDescriptors = []RouteDescriptor{
 								Name:        "Invalid Name or Tag",
 								Description: "The specified `name` or `tag` were invalid and the delete was unable to proceed.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeTagInvalid,
 								},
@@ -680,7 +657,7 @@ var routeDescriptors = []RouteDescriptor{
 										Format:      "<length>",
 									},
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeUnauthorized,
 								},
 								Body: BodyDescriptor{
@@ -692,7 +669,7 @@ var routeDescriptors = []RouteDescriptor{
 								Name:        "Unknown Manifest",
 								Description: "The specified `name` or `tag` are unknown to the registry and the delete was unable to proceed. Clients can assume the manifest was already deleted if this response is returned.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameUnknown,
 									ErrorCodeManifestUnknown,
 								},
@@ -765,7 +742,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was a problem with the request that needs to be addressed by the client, such as an invalid `name` or `tag`.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeDigestInvalid,
 								},
@@ -782,7 +759,7 @@ var routeDescriptors = []RouteDescriptor{
 									ContentType: "application/json; charset=utf-8",
 									Format:      errorsBody,
 								},
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameUnknown,
 									ErrorCodeBlobUnknown,
 								},
@@ -834,7 +811,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was a problem with the request that needs to be addressed by the client, such as an invalid `name` or `tag`.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeDigestInvalid,
 								},
@@ -846,7 +823,7 @@ var routeDescriptors = []RouteDescriptor{
 							unauthorizedResponse,
 							{
 								StatusCode: http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameUnknown,
 									ErrorCodeBlobUnknown,
 								},
@@ -926,7 +903,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Name:       "Invalid Name or Digest",
 								StatusCode: http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 								},
@@ -970,7 +947,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Name:       "Invalid Name or Digest",
 								StatusCode: http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 								},
@@ -1024,7 +1001,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was an error processing the upload and it must be restarted.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 									ErrorCodeBlobUploadInvalid,
@@ -1038,7 +1015,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The upload is unknown to the registry. The upload must be restarted.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUploadUnknown,
 								},
 								Body: BodyDescriptor{
@@ -1096,7 +1073,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was an error processing the upload and it must be restarted.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 									ErrorCodeBlobUploadInvalid,
@@ -1110,7 +1087,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The upload is unknown to the registry. The upload must be restarted.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUploadUnknown,
 								},
 								Body: BodyDescriptor{
@@ -1175,7 +1152,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was an error processing the upload and it must be restarted.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 									ErrorCodeBlobUploadInvalid,
@@ -1189,7 +1166,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The upload is unknown to the registry. The upload must be restarted.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUploadUnknown,
 								},
 								Body: BodyDescriptor{
@@ -1266,7 +1243,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "There was an error processing the upload and it must be restarted.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeDigestInvalid,
 									ErrorCodeNameInvalid,
 									ErrorCodeBlobUploadInvalid,
@@ -1280,7 +1257,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The upload is unknown to the registry. The upload must be restarted.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUploadUnknown,
 								},
 								Body: BodyDescriptor{
@@ -1321,7 +1298,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "An error was encountered processing the delete. The client may ignore this error.",
 								StatusCode:  http.StatusBadRequest,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeNameInvalid,
 									ErrorCodeBlobUploadInvalid,
 								},
@@ -1334,7 +1311,7 @@ var routeDescriptors = []RouteDescriptor{
 							{
 								Description: "The upload is unknown to the registry. The client may ignore this error and assume the upload has been deleted.",
 								StatusCode:  http.StatusNotFound,
-								ErrorCodes: []ErrorCode{
+								ErrorCodes: []errcode.ErrorCode{
 									ErrorCodeBlobUploadUnknown,
 								},
 								Body: BodyDescriptor{
@@ -1350,143 +1327,11 @@ var routeDescriptors = []RouteDescriptor{
 	},
 }
 
-// ErrorDescriptors provides a list of HTTP API Error codes that may be
-// encountered when interacting with the registry API.
-var errorDescriptors = []ErrorDescriptor{
-	{
-		Code:    ErrorCodeUnknown,
-		Value:   "UNKNOWN",
-		Message: "unknown error",
-		Description: `Generic error returned when the error does not have an
-		API classification.`,
-	},
-	{
-		Code:    ErrorCodeUnsupported,
-		Value:   "UNSUPPORTED",
-		Message: "The operation is unsupported.",
-		Description: `The operation was unsupported due to a missing
-		implementation or invalid set of parameters.`,
-	},
-	{
-		Code:    ErrorCodeUnauthorized,
-		Value:   "UNAUTHORIZED",
-		Message: "access to the requested resource is not authorized",
-		Description: `The access controller denied access for the operation on
-		a resource. Often this will be accompanied by a 401 Unauthorized
-		response status.`,
-	},
-	{
-		Code:    ErrorCodeDigestInvalid,
-		Value:   "DIGEST_INVALID",
-		Message: "provided digest did not match uploaded content",
-		Description: `When a blob is uploaded, the registry will check that
-		the content matches the digest provided by the client. The error may
-		include a detail structure with the key "digest", including the
-		invalid digest string. This error may also be returned when a manifest
-		includes an invalid layer digest.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest, http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeSizeInvalid,
-		Value:   "SIZE_INVALID",
-		Message: "provided length did not match content length",
-		Description: `When a layer is uploaded, the provided size will be
-		checked against the uploaded content. If they do not match, this error
-		will be returned.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest},
-	},
-	{
-		Code:    ErrorCodeNameInvalid,
-		Value:   "NAME_INVALID",
-		Message: "invalid repository name",
-		Description: `Invalid repository name encountered either during
-		manifest validation or any API operation.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest, http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeTagInvalid,
-		Value:   "TAG_INVALID",
-		Message: "manifest tag did not match URI",
-		Description: `During a manifest upload, if the tag in the manifest
-		does not match the uri tag, this error will be returned.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest, http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeNameUnknown,
-		Value:   "NAME_UNKNOWN",
-		Message: "repository name not known to registry",
-		Description: `This is returned if the name used during an operation is
-		unknown to the registry.`,
-		HTTPStatusCodes: []int{http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeManifestUnknown,
-		Value:   "MANIFEST_UNKNOWN",
-		Message: "manifest unknown",
-		Description: `This error is returned when the manifest, identified by
-		name and tag is unknown to the repository.`,
-		HTTPStatusCodes: []int{http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeManifestInvalid,
-		Value:   "MANIFEST_INVALID",
-		Message: "manifest invalid",
-		Description: `During upload, manifests undergo several checks ensuring
-		validity. If those checks fail, this error may be returned, unless a
-		more specific error is included. The detail will contain information
-		the failed validation.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest},
-	},
-	{
-		Code:    ErrorCodeManifestUnverified,
-		Value:   "MANIFEST_UNVERIFIED",
-		Message: "manifest failed signature verification",
-		Description: `During manifest upload, if the manifest fails signature
-		verification, this error will be returned.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest},
-	},
-	{
-		Code:    ErrorCodeBlobUnknown,
-		Value:   "BLOB_UNKNOWN",
-		Message: "blob unknown to registry",
-		Description: `This error may be returned when a blob is unknown to the
-		registry in a specified repository. This can be returned with a
-		standard get or if a manifest references an unknown layer during
-		upload.`,
-		HTTPStatusCodes: []int{http.StatusBadRequest, http.StatusNotFound},
-	},
-
-	{
-		Code:    ErrorCodeBlobUploadUnknown,
-		Value:   "BLOB_UPLOAD_UNKNOWN",
-		Message: "blob upload unknown to registry",
-		Description: `If a blob upload has been cancelled or was never
-		started, this error code may be returned.`,
-		HTTPStatusCodes: []int{http.StatusNotFound},
-	},
-	{
-		Code:    ErrorCodeBlobUploadInvalid,
-		Value:   "BLOB_UPLOAD_INVALID",
-		Message: "blob upload invalid",
-		Description: `The blob upload encountered an error and can no
-		longer proceed.`,
-		HTTPStatusCodes: []int{http.StatusNotFound},
-	},
-}
-
-var errorCodeToDescriptors map[ErrorCode]ErrorDescriptor
-var idToDescriptors map[string]ErrorDescriptor
 var routeDescriptorsMap map[string]RouteDescriptor
 
 func init() {
-	errorCodeToDescriptors = make(map[ErrorCode]ErrorDescriptor, len(errorDescriptors))
-	idToDescriptors = make(map[string]ErrorDescriptor, len(errorDescriptors))
 	routeDescriptorsMap = make(map[string]RouteDescriptor, len(routeDescriptors))
 
-	for _, descriptor := range errorDescriptors {
-		errorCodeToDescriptors[descriptor.Code] = descriptor
-		idToDescriptors[descriptor.Value] = descriptor
-	}
 	for _, descriptor := range routeDescriptors {
 		routeDescriptorsMap[descriptor.Name] = descriptor
 	}
