+++ registry/handlers/blobupload.go
@@ -10,6 +10,7 @@ import (
 	"github.com/docker/distribution"
 	ctxu "github.com/docker/distribution/context"
 	"github.com/docker/distribution/digest"
+	"github.com/docker/distribution/registry/api/errcode"
 	"github.com/docker/distribution/registry/api/v2"
 	"github.com/gorilla/handlers"
 )
@@ -36,7 +37,6 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 		if err != nil {
 			return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 				ctxu.GetLogger(ctx).Infof("error resolving upload: %v", err)
-				w.WriteHeader(http.StatusBadRequest)
 				buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 			})
 		}
@@ -45,7 +45,6 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 		if state.Name != ctx.Repository.Name() {
 			return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 				ctxu.GetLogger(ctx).Infof("mismatched repository name in upload state: %q != %q", state.Name, buh.Repository.Name())
-				w.WriteHeader(http.StatusBadRequest)
 				buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 			})
 		}
@@ -53,7 +52,6 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 		if state.UUID != buh.UUID {
 			return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 				ctxu.GetLogger(ctx).Infof("mismatched uuid in upload state: %q != %q", state.UUID, buh.UUID)
-				w.WriteHeader(http.StatusBadRequest)
 				buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 			})
 		}
@@ -64,14 +62,12 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 			ctxu.GetLogger(ctx).Errorf("error resolving upload: %v", err)
 			if err == distribution.ErrBlobUploadUnknown {
 				return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(http.StatusNotFound)
 					buh.Errors.Push(v2.ErrorCodeBlobUploadUnknown, err)
 				})
 			}
 
 			return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(http.StatusInternalServerError)
-				buh.Errors.Push(v2.ErrorCodeUnknown, err)
+				buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 			})
 		}
 		buh.Upload = upload
@@ -85,7 +81,6 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 				defer upload.Close()
 				ctxu.GetLogger(ctx).Infof("error seeking blob upload: %v", err)
 				return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(http.StatusBadRequest)
 					buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 					upload.Cancel(buh)
 				})
@@ -93,7 +88,6 @@ func blobUploadDispatcher(ctx *Context, r *http.Request) http.Handler {
 				defer upload.Close()
 				ctxu.GetLogger(ctx).Infof("seek to wrong offest: %d != %d", nn, buh.State.Offset)
 				return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(http.StatusBadRequest)
 					buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 					upload.Cancel(buh)
 				})
@@ -125,8 +119,7 @@ func (buh *blobUploadHandler) StartBlobUpload(w http.ResponseWriter, r *http.Req
 	blobs := buh.Repository.Blobs(buh)
 	upload, err := blobs.Create(buh)
 	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError) // Error conditions here?
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -134,8 +127,7 @@ func (buh *blobUploadHandler) StartBlobUpload(w http.ResponseWriter, r *http.Req
 	defer buh.Upload.Close()
 
 	if err := buh.blobUploadResponse(w, r, true); err != nil {
-		w.WriteHeader(http.StatusInternalServerError) // Error conditions here?
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -146,7 +138,6 @@ func (buh *blobUploadHandler) StartBlobUpload(w http.ResponseWriter, r *http.Req
 // GetUploadStatus returns the status of a given upload, identified by id.
 func (buh *blobUploadHandler) GetUploadStatus(w http.ResponseWriter, r *http.Request) {
 	if buh.Upload == nil {
-		w.WriteHeader(http.StatusNotFound)
 		buh.Errors.Push(v2.ErrorCodeBlobUploadUnknown)
 		return
 	}
@@ -155,8 +146,7 @@ func (buh *blobUploadHandler) GetUploadStatus(w http.ResponseWriter, r *http.Req
 	// resumable upload is supported. This will enable returning a non-zero
 	// range for clients to begin uploading at an offset.
 	if err := buh.blobUploadResponse(w, r, true); err != nil {
-		w.WriteHeader(http.StatusInternalServerError) // Error conditions here?
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -167,14 +157,13 @@ func (buh *blobUploadHandler) GetUploadStatus(w http.ResponseWriter, r *http.Req
 // PatchBlobData writes data to an upload.
 func (buh *blobUploadHandler) PatchBlobData(w http.ResponseWriter, r *http.Request) {
 	if buh.Upload == nil {
-		w.WriteHeader(http.StatusNotFound)
 		buh.Errors.Push(v2.ErrorCodeBlobUploadUnknown)
 		return
 	}
 
 	ct := r.Header.Get("Content-Type")
 	if ct != "" && ct != "application/octet-stream" {
-		w.WriteHeader(http.StatusBadRequest)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, fmt.Errorf("Bad Content-Type"))
 		// TODO(dmcgowan): encode error
 		return
 	}
@@ -184,14 +173,12 @@ func (buh *blobUploadHandler) PatchBlobData(w http.ResponseWriter, r *http.Reque
 	// Copy the data
 	if _, err := io.Copy(buh.Upload, r.Body); err != nil {
 		ctxu.GetLogger(buh).Errorf("unknown error copying into upload: %v", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
 	if err := buh.blobUploadResponse(w, r, false); err != nil {
-		w.WriteHeader(http.StatusInternalServerError) // Error conditions here?
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -205,7 +192,6 @@ func (buh *blobUploadHandler) PatchBlobData(w http.ResponseWriter, r *http.Reque
 // url of the blob.
 func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *http.Request) {
 	if buh.Upload == nil {
-		w.WriteHeader(http.StatusNotFound)
 		buh.Errors.Push(v2.ErrorCodeBlobUploadUnknown)
 		return
 	}
@@ -214,7 +200,6 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 
 	if dgstStr == "" {
 		// no digest? return error, but allow retry.
-		w.WriteHeader(http.StatusBadRequest)
 		buh.Errors.Push(v2.ErrorCodeDigestInvalid, "digest missing")
 		return
 	}
@@ -222,7 +207,6 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 	dgst, err := digest.ParseDigest(dgstStr)
 	if err != nil {
 		// no digest? return error, but allow retry.
-		w.WriteHeader(http.StatusNotFound)
 		buh.Errors.Push(v2.ErrorCodeDigestInvalid, "digest parsing failed")
 		return
 	}
@@ -230,8 +214,7 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 	// Read in the data, if any.
 	if _, err := io.Copy(buh.Upload, r.Body); err != nil {
 		ctxu.GetLogger(buh).Errorf("unknown error copying into upload: %v", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -246,17 +229,14 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 	if err != nil {
 		switch err := err.(type) {
 		case distribution.ErrBlobInvalidDigest:
-			w.WriteHeader(http.StatusBadRequest)
 			buh.Errors.Push(v2.ErrorCodeDigestInvalid, err)
 		default:
 			switch err {
 			case distribution.ErrBlobInvalidLength, distribution.ErrBlobDigestUnsupported:
-				w.WriteHeader(http.StatusBadRequest)
 				buh.Errors.Push(v2.ErrorCodeBlobUploadInvalid, err)
 			default:
 				ctxu.GetLogger(buh).Errorf("unknown error completing upload: %#v", err)
-				w.WriteHeader(http.StatusInternalServerError)
-				buh.Errors.Push(v2.ErrorCodeUnknown, err)
+				buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 			}
 
 		}
@@ -273,8 +253,7 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 	// Build our canonical blob url
 	blobURL, err := buh.urlBuilder.BuildBlobURL(buh.Repository.Name(), desc.Digest)
 	if err != nil {
-		buh.Errors.Push(v2.ErrorCodeUnknown, err)
-		w.WriteHeader(http.StatusInternalServerError)
+		buh.Errors.Push(errcode.ErrorCodeUnknown, err)
 		return
 	}
 
@@ -287,7 +266,6 @@ func (buh *blobUploadHandler) PutBlobUploadComplete(w http.ResponseWriter, r *ht
 // CancelBlobUpload cancels an in-progress upload of a blob.
 func (buh *blobUploadHandler) CancelBlobUpload(w http.ResponseWriter, r *http.Request) {
 	if buh.Upload == nil {
-		w.WriteHeader(http.StatusNotFound)
 		buh.Errors.Push(v2.ErrorCodeBlobUploadUnknown)
 		return
 	}
@@ -295,7 +273,6 @@ func (buh *blobUploadHandler) CancelBlobUpload(w http.ResponseWriter, r *http.Re
 	w.Header().Set("Docker-Upload-UUID", buh.UUID)
 	if err := buh.Upload.Cancel(buh); err != nil {
 		ctxu.GetLogger(buh).Errorf("error encountered canceling upload: %v", err)
-		w.WriteHeader(http.StatusInternalServerError)
 		buh.Errors.PushErr(err)
 	}
